+++ src/Request.php
@@ -3,11 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\Http;
 
-use Innmind\Http\{
-    Message\Method,
-    ProtocolVersion,
-    Headers,
-};
 use Innmind\Url\Url;
 use Innmind\Filesystem\File\Content;